% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amino_acid_calc.R
\name{count_carbon_atoms}
\alias{count_carbon_atoms}
\title{count_carbon_atoms}
\usage{
count_carbon_atoms(sequence)
}
\arguments{
\item{sequence}{Alphabetical sequence of peptides}
}
\description{
Define a function that takes a peptide name and outputs the number of C elements it contains.
}
\examples{
count_carbon_atoms("ACDE")
}
