% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eics.R
\name{generate_points_as_EICs}
\alias{generate_points_as_EICs}
\title{Generating points forming EIC}
\usage{
generate_points_as_EICs(raw_data, ranges)
}
\arguments{
\item{raw_data}{raw_data}

\item{ranges}{ranges}
}
\description{
Define a function that takes a mzML data and rt/mz ranges.
}
