% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amino_acid_calc.R
\name{generate_rt_mz_ranges}
\alias{generate_rt_mz_ranges}
\title{Generating ranges}
\usage{
generate_rt_mz_ranges(sequence, rt, mz, charge, rt_tolerance, mz_tolerance)
}
\arguments{
\item{sequence}{Alphabetical sequence of peptides}

\item{rt}{rt}

\item{mz}{mz}

\item{charge}{charge}

\item{rt_tolerance}{rt_tolerance}

\item{mz_tolerance}{mz_tolerance}
}
\description{
Define a function that takes a peptide name and outputs the ranges of rt and mz.
}
\examples{
count_carbon_atoms("ACDE")
}
