% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amino_acid_calc.R
\name{sum_amino_acids}
\alias{sum_amino_acids}
\title{sum_amino_acids}
\usage{
sum_amino_acids(sequence)
}
\arguments{
\item{sequence}{Alphabetical sequence of peptides}
}
\description{
Define a function that inputs a string and accumulates the number of digits based on each letter.
}
\examples{
sum_amino_acids("ACDE")
}
